requirements.txt
# HEPTAPOD Requirements
# Install with: pip install -r requirements.txt

# Core Dependencies
# ================

# Orchestral AI Framework (LLM orchestration)
orchestral-ai>=1.2.1

# LLM Provider SDKs
anthropic>=0.75.0        # Claude API
openai>=2.13.0           # GPT API
google-generativeai>=0.8.6  # Gemini API (optional)
ollama>=0.6.1            # Local models (optional)
groq>=1.0.0              # Groq API (optional)
mistralai>=1.10.0        # Mistral API (optional)

# Scientific Computing
numpy>=2.3.5
pandas>=2.3.3
scipy>=1.16.3

# Physics-Specific Packages
# ==========================
pythia8mc>=8.316.0       # Pythia8 Python bindings (requires Python 3.12 or 3.13, NOT 3.14+)
sherpa-mc>=3.0.3b1       # Sherpa event generator (requires Python 3.12 or higher)
pylhe>=1.0.2             # LHE file parsing
particle>=0.26.0         # Particle data and PDG IDs
hepunits>=2.4.3          # HEP unit conversions
awkward>=2.8.11          # Awkward arrays for HEP data
awkward-cpp>=51          # C++ backend for awkward
vector>=1.7.0            # Vector operations for HEP

# Data Validation
pydantic>=2.12.5         # Type validation and schema enforcement

# Terminal UI
rich>=14.2.0             # Rich terminal output

# Web Server (for app.server)
fastapi>=0.125.0         # Web framework
uvicorn>=0.38.0          # ASGI server
websockets>=15.0.1       # WebSocket support
starlette>=0.50.0        # ASGI toolkit

# Visualization
matplotlib>=3.10.8       # Plotting
graphviz>=0.21           # Graph visualization

# Utilities
python-dotenv>=1.2.1     # Environment variable management
pyyaml>=6.0.3            # YAML parsing
invoke>=2.2.1            # Task execution

# Development Tools (Optional)
# =============================
# Uncomment these for development:
# pytest>=7.4.0          # Testing framework
# black>=23.0.0          # Code formatting
# flake8>=6.0.0          # Linting
# mypy>=1.0.0            # Type checking
